requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
drf-yasg>=1.21.7
python-decouple>=3.8
psycopg2-binary>=2.9.9

